requirements.txt
aiohttp
aiohttp-apispec
aiohttp-session
aioresponses
aiosignal
alembic
apispec
async-timeout
asyncpg
attrs
cffi
charset-normalizer
coverage
cryptography
freezegun
frozenlist
greenlet
idna
iniconfig
Jinja2
Mako
MarkupSafe
marshmallow
multidict
packaging
pluggy
py
pycparser
pyparsing
pytest
pytest-aiohttp
pytest-asyncio
pytest-cov
pytest-mock
python-dateutil
PyYAML
six
SQLAlchemy
tomli
webargs
yarl